Change type of price fields for better accuracy

+++ src/main/java/ru/tersoft/ticketsale/entity/Order.java
@@ -9,6 +9,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
@@ -42,7 +43,7 @@ public class Order implements Serializable {
 
     @Column(name = "total", columnDefinition = "decimal(19,4)")
     @ApiModelProperty(example = "999.99")
-    private Float total;
+    private BigDecimal total;
 
     @Column(name = "payed")
     private Boolean payed;
@@ -71,11 +72,11 @@ public class Order implements Serializable {
         this.orderdate = orderdate;
     }
 
-    public Float getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Float total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
